Moved Lap.java into Task.java, Fixed loading
## Lap.java ##
Lap.java's important elements were moved into Task.java in order to fix
the crash with loading as GSON was not liking loading an array with a
class nested in it.

## Task.java ##
+gained hours, minutes, seconds, and millis variables from Lap.java
+moved all neccisary functions over from Lap.java (all gets and
incriments)
+changed any function that interacted with Lap.java to instead pull from
the classes local variables.
+changed all variables to private with their own get functions.
+added setting time related variables in constrcutor.

## MainWindow.java ##
+moved JTabbedPane taskTabs declaration outside of initalise so that
load can use it properly to generate the tabs on loading.
+removed functionality from what was the "New Task" button on each tab
temporarily.
+ changed "New Task" button into "Delete Task" on each tab so there will
be a way to remove finished tasks.
+fixed that addTab was always trying to use the first tabs name for the
task name.
+moved loading of file into its own seperate function.
+fixed file loading by changing what GSON is expecting from an ArrayList
of strings to an ArrayList of our custon Task class.
+updated useage of addTab in file loading to work with the new addTab
function.
+Load properly closes the bufferedReader it uses now so loading doesnt
hang for a few seconds on launch.
+a bunch of comment cleanup and removal, including temporary code while
designing the dynamicaly generated tabs layouts.
+ changed saveing to use gson.toJson(<variable>, <fileWriter>) in stead
of having the filreWriter writting the string gson generates itself.